All text moved to view
Text loading for CLI moved to a file specifically built for view.
